start of middleware test
