+++ websocket/client_test.go
@@ -4,73 +4,16 @@ import (
 	"encoding/json"
 	"errors"
 	"net/http"
-	"net/url"
 	"testing"
 
 	gorilla "github.com/gorilla/websocket"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/suite"
 
 	"github.com/thecoderstudio/apollo-agent/oauth"
+	"github.com/thecoderstudio/apollo-agent/testutil"
 	"github.com/thecoderstudio/apollo-agent/websocket"
 )
-
-var u = url.URL{Scheme: "ws", Host: "localhost:8000", Path: "/ws"}
-
-type ConnMock struct {
-	mock.Mock
-}
-
-func (mocked *ConnMock) Close() error {
-	args := mocked.Called()
-	return args.Error(0)
-}
-
-func (mocked *ConnMock) ReadMessage() (messageType int, p []byte, err error) {
-	args := mocked.Called()
-
-	var message []byte
-	if args.Get(1) != nil {
-		message = args.Get(1).([]byte)
-	}
-
-	return args.Int(0), message, args.Error(2)
-}
-
-func (mocked *ConnMock) WriteMessage(messageType int, data []byte) error {
-	args := mocked.Called(messageType, data)
-	return args.Error(0)
-}
-
-func (mocked *ConnMock) MockClosed(expectedError error) {
-	mocked.On("Close").Return(nil)
-	mocked.On(
-		"WriteMessage",
-		gorilla.CloseMessage,
-		gorilla.FormatCloseMessage(gorilla.CloseNormalClosure, ""),
-	).Return(expectedError)
-}
-
-type DialerMock struct {
-	mock.Mock
-}
-
-func (mocked *DialerMock) Dial(urlString string, header http.Header) (websocket.Connection, *http.Response, error) {
-	args := mocked.Called(urlString, header)
-
-	var connection websocket.Connection
-	if args.Get(0) != nil {
-		connection = args.Get(0).(websocket.Connection)
-	}
-
-	var response http.Response
-	if args.Get(1) != nil {
-		response = args.Get(1).(http.Response)
-	}
-
-	return connection, &response, args.Error(2)
-}
 
 type ClientTestSuite struct {
 	suite.Suite
@@ -81,13 +24,13 @@ func (suite *ClientTestSuite) TestListenForShellIOSuccess() {
 	in := make(chan websocket.ShellIO)
 	defer close(in)
 
-	mockConn := new(ConnMock)
+	mockConn := new(testutil.ConnMock)
 	mockConn.MockClosed(nil)
 	mockConn.On("ReadMessage").Return(0, []byte("{\"command\": \"new connection\"}"), nil).Once()
 	mockConn.On("ReadMessage").Return(0, []byte("{\"message\": \"test message\"}"), nil)
 
-	wsClient := createWsClient(mockConn)
-	done := wsClient.Listen(u, oauth.AccessToken{}, &in, &interrupt)
+	wsClient := testutil.CreateWsClient(mockConn)
+	done := wsClient.Listen(testutil.Url, oauth.AccessToken{}, &in, &interrupt)
 	command := <-wsClient.Commands()
 	message := <-wsClient.Out()
 
@@ -106,12 +49,12 @@ func (suite *ClientTestSuite) TestCloseConnectionWriteError() {
 	in := make(chan websocket.ShellIO)
 	defer close(in)
 
-	mockConn := new(ConnMock)
+	mockConn := new(testutil.ConnMock)
 	mockConn.MockClosed(expectedError)
 	mockConn.On("ReadMessage").Maybe().Return(0, nil, nil)
 
-	wsClient := createWsClient(mockConn)
-	done := wsClient.Listen(u, oauth.AccessToken{}, &in, &interrupt)
+	wsClient := testutil.CreateWsClient(mockConn)
+	done := wsClient.Listen(testutil.Url, oauth.AccessToken{}, &in, &interrupt)
 
 	close(interrupt)
 
@@ -126,11 +69,11 @@ func (suite *ClientTestSuite) TestConnectionError() {
 	defer close(interrupt)
 	defer close(in)
 
-	mockDialer := new(DialerMock)
-	mockDialer.On("Dial", u.String(), http.Header{"Authorization": []string{" "}}).Return(nil, nil, expectedError)
+	mockDialer := new(testutil.DialerMock)
+	mockDialer.On("Dial", testutil.Url.String(), http.Header{"Authorization": []string{" "}}).Return(nil, nil, expectedError)
 
 	wsClient := websocket.CreateClient(mockDialer)
-	wsClient.Listen(u, oauth.AccessToken{}, &in, &interrupt)
+	wsClient.Listen(testutil.Url, oauth.AccessToken{}, &in, &interrupt)
 	err := <-wsClient.Errs()
 
 	mockDialer.AssertExpectations(suite.T())
@@ -144,12 +87,12 @@ func (suite *ClientTestSuite) TestReadMessageError() {
 	defer close(interrupt)
 	defer close(in)
 
-	mockConn := new(ConnMock)
+	mockConn := new(testutil.ConnMock)
 	mockConn.On("Close").Return(nil)
 	mockConn.On("ReadMessage").Return(0, nil, expectedError)
 
-	wsClient := createWsClient(mockConn)
-	done := wsClient.Listen(u, oauth.AccessToken{}, &in, &interrupt)
+	wsClient := testutil.CreateWsClient(mockConn)
+	done := wsClient.Listen(testutil.Url, oauth.AccessToken{}, &in, &interrupt)
 	err := <-wsClient.Errs()
 
 	assert.NotNil(suite.T(), <-done)
@@ -168,7 +111,7 @@ func (suite *ClientTestSuite) TestWriteMessage() {
 	}
 	jsonShellIO, _ := json.Marshal(testShellIO)
 
-	mockConn := new(ConnMock)
+	mockConn := new(testutil.ConnMock)
 	mockConn.MockClosed(nil)
 	mockConn.On("ReadMessage").Return(0, nil, nil)
 	mockConn.On(
@@ -176,8 +119,8 @@ func (suite *ClientTestSuite) TestWriteMessage() {
 		gorilla.TextMessage,
 		jsonShellIO).Return(nil)
 
-	wsClient := createWsClient(mockConn)
-	done := wsClient.Listen(u, oauth.AccessToken{}, &in, &interrupt)
+	wsClient := testutil.CreateWsClient(mockConn)
+	done := wsClient.Listen(testutil.Url, oauth.AccessToken{}, &in, &interrupt)
 
 	in <- testShellIO
 
@@ -190,12 +133,4 @@ func (suite *ClientTestSuite) TestWriteMessage() {
 
 func TestClientSuite(t *testing.T) {
 	suite.Run(t, new(ClientTestSuite))
-}
-
-func createWsClient(mockConn *ConnMock) websocket.Client {
-	mockDialer := new(DialerMock)
-	mockDialer.On("Dial", u.String(), http.Header{"Authorization": []string{" "}}).Return(mockConn, nil, nil)
-
-	wsClient := websocket.CreateClient(mockDialer)
-	return wsClient
 }
